refactor: extract technician calendar readonly routes
